corrected

+++ Class creation/Class creation/classA.cs
@@ -8,7 +8,7 @@ namespace Class_creation
 {
      class ClassA
     {
-        public static double Hourly(double num1, double num2)
+        public static int Hourly(int num1, int num2)
         {
 
             return num1 / num2;
@@ -18,13 +18,13 @@ namespace Class_creation
 
         }
 
-        public static double Taxes(double num1, double num2)
+        public static int Taxes(int num1, int num2)
         {
             return num1 * num2;
 
         }
 
-        public static double Lifetime(double num1, double num2)
+        public static int Lifetime(int num1, int num2)
         {
             return num1 * num2;
